requirements.txt
streamlit
pdfplumber
python-docx
pandas
altair
pytest
reportlab
weasyprint
black
ruff
isort
streamlit
pdfplumber
python-docx
pandas
altair
pytest
